Use data counter instead of repeat count for sample delay
This allows for a 32-bit delay to be specified instead of an 8-bit one.

+++ acquisition.h
@@ -1,1 +1,1 @@
-extern void start_acquisition(uint8_t bits, uint8_t delay, uint16_t clock_divisor_x2);
+extern void start_acquisition(uint8_t bits, uint32_t delay, uint16_t clock_divisor_x2);
